refactor: jwt 토큰 관련 파일 이동 및 불필요한 주석 삭제
