+++ src/main/java/postman/bottler/user/auth/CustomUserDetailsService.java
@@ -1,4 +1,4 @@
-package postman.bottler.user.auth.service;
+package postman.bottler.user.auth;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
