+++ src/main/java/postman/bottler/user/auth/JwtTokenProvider.java
@@ -2,7 +2,6 @@ package postman.bottler.user.auth;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.ExpiredJwtException;
-import io.jsonwebtoken.Jws;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.MalformedJwtException;
 import io.jsonwebtoken.SignatureAlgorithm;
@@ -13,6 +12,7 @@ import java.security.Key;
 import java.util.Date;
 import java.util.List;
 import java.util.stream.Collectors;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.InitializingBean;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
@@ -21,9 +21,9 @@ import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Component;
-import postman.bottler.user.auth.service.CustomUserDetailsService;
 
 @Component
+@Slf4j
 public class JwtTokenProvider implements InitializingBean {
 
     private final String secretKey; // JWT 비밀키
@@ -53,20 +53,16 @@ public class JwtTokenProvider implements InitializingBean {
     }
 
     public String createToken(Authentication authentication) {
-        // 사용자 이름 (주로 username 또는 userId)
         String username = authentication.getName();
 
-        // 권한 정보를 문자열로 변환
         String authorities = authentication.getAuthorities().stream()
                 .map(GrantedAuthority
                         ::getAuthority)
                 .collect(Collectors.joining(","));
 
-        // 현재 시간과 만료 시간 설정
         Date now = new Date();
         Date expiryDate = new Date(now.getTime() + accessTokenExpirationTime);
 
-        // JWT 생성 및 서명
         return Jwts.builder()
                 .setSubject(username) // 사용자 정보 설정
                 .claim(AUTHORITIES_KEY, authorities) // 권한 정보 추가
@@ -78,56 +74,41 @@ public class JwtTokenProvider implements InitializingBean {
 
 
     public Authentication getAuthentication(String token) {
-        // JWT 토큰에서 클레임(Claims) 추출
         Claims claims = Jwts.parserBuilder()
-                .setSigningKey(key) // 서명 키 설정
+                .setSigningKey(key)
                 .build()
                 .parseClaimsJws(token)
                 .getBody();
 
-        // 사용자 이름 또는 ID 추출
         String username = claims.getSubject();
 
-        // 권한 정보 추출 및 변환
         List<SimpleGrantedAuthority> authorities =
-                ((List<?>) claims.get(AUTHORITIES_KEY)) // 권한 키에서 리스트 추출
+                ((List<?>) claims.get(AUTHORITIES_KEY))
                         .stream()
-                        .map(authority -> new SimpleGrantedAuthority(authority.toString())) // SimpleGrantedAuthority로 변환
+                        .map(authority -> new SimpleGrantedAuthority(authority.toString()))
                         .collect(Collectors.toList());
 
-        // UserDetailsService에서 사용자 정보를 로드하거나, 간단히 UserDetails 생성
         UserDetails userDetails = customUserDetailsService.loadUserByUsername(username);
 
-        // Authentication 객체 반환 (UsernamePasswordAuthenticationToken 사용)
         return new UsernamePasswordAuthenticationToken(userDetails, token, authorities);
     }
 
     public boolean validateToken(String token) {
         try {
-            // JWT 토큰 파싱 및 유효성 검증
-            Jws<Claims> claimsJws = Jwts.parserBuilder()
+            Jwts.parserBuilder()
                     .setSigningKey(key) // 서명 키 설정
                     .build()
                     .parseClaimsJws(token);
-
-            // 추가 검증 로직 (선택적)
-            // claimsJws.getBody()를 사용하여 페이로드 내용을 확인할 수 있음
-
-            return true; // 유효한 토큰
+            return true;
         } catch (SignatureException e) {
-            // 서명 불일치 예외 처리
-            System.out.println("Invalid JWT signature");
+            log.error("jwt signature가 불일치합니다. " + e);
         } catch (MalformedJwtException e) {
-            // 구조가 잘못된 JWT 처리
-            System.out.println("Invalid JWT token");
+            log.error("유효하지 않은 jwt 토큰입니다. " + e);
         } catch (ExpiredJwtException e) {
-            // 만료된 JWT 처리
-            System.out.println("Expired JWT token");
+            log.error("만료된 jwt 토큰입니다. " + e);
         } catch (IllegalArgumentException e) {
-            // 클레임이 비어있는 경우 처리
-            System.out.println("JWT claims string is empty");
+            log.error("jwt claims 비어있습니다. " + e);
         }
-
-        return false; // 유효하지 않은 토큰
+        return false;
     }
 }
